requirements.txt
openai>=1.0.0
youtube-transcript-api
python-dotenv
